Fix: Separate DTO package for DB interactions
