+++ app/src/main/kotlin/org/example/actions/NotificationAction.kt
@@ -18,7 +18,11 @@ import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toInstant
 import kotlinx.datetime.toLocalDateTime
-import kotlinx.serialization.Serializable
+import org.example.dto.DbCreateRequest
+import org.example.dto.DbNotificationRow
+import org.example.dto.DbReadRequest
+import org.example.dto.DbResponse
+import org.example.dto.DbUpdateRequest
 import org.example.model.Notification
 import org.example.model.NotificationActionEnum
 import org.example.model.NotificationSnoozeEnum
@@ -43,86 +47,6 @@ class NotificationAction(private val config: ApplicationConfig) : INotificationA
       }
 
   private val httpClient = HttpClient { install(ContentNegotiation) { json() } }
-
-  /**
-   * Запрос на создание записи в БД
-   *
-   * @property table Название таблицы
-   * @property data Данные, которые нужно вставить
-   */
-  @Serializable data class DbCreateRequest(val table: String, val data: Map<String, String>)
-
-  /**
-   * Запрос на чтение данных из БД
-   *
-   * @property table Название таблицы
-   * @property columns Список столбцов, которые необходимо прочитать, в данном случае читаем все
-   * @property filters Фильтры для запроса
-   */
-  @Serializable
-  data class DbReadRequest(
-      val table: String,
-      val columns: List<String> = listOf("*"),
-      val filters: Map<String, String>? = null
-  )
-
-  /**
-   * Запрос на обновление данных в БД.
-   *
-   * @property table Название таблицы
-   * @property data Новые значения
-   * @property condition Условие обновления
-   * @property conditionParams Параметры для условия
-   */
-  @Serializable
-  data class DbUpdateRequest(
-      val table: String,
-      val data: Map<String, String>,
-      val condition: String,
-      val conditionParams: List<String>
-  )
-
-  /**
-   * Запрос на удаление данных в БД.
-   *
-   * @property table Название таблицы
-   * @property condition Условие удаления
-   * @property conditionParams Параметры условия
-   */
-  @Serializable
-  data class DbDeleteRequest(
-      val table: String,
-      val condition: String,
-      val conditionParams: List<String>
-  )
-
-  /**
-   * Ответ от БД
-   *
-   * @property success Успех операции
-   * @property error Сообщение об ошибке
-   */
-  @Serializable data class DbResponse(val success: Boolean? = null, val error: String? = null)
-
-  /**
-   * Строка таблицы уведомлений
-   *
-   * @property id ID уведомления
-   * @property userid ID пользователя
-   * @property title Заголовок уведомления
-   * @property description Описание уведомления
-   * @property create_date Дата создания уведомления
-   * @property notification_date Дата уведомления
-   */
-  @Serializable
-  data class DbNotificationRow(
-      val id: String,
-      val userid: String,
-      val title: String,
-      val description: String,
-      val create_date: String,
-      val notification_date: String
-  )
 
   /**
    * Создает новое уведомление
